requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.9.4
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
asteroid-filterbanks==0.4.0
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
backports.cached-property==1.0.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cmaes==0.9.1
colorama==0.4.6
colorlog==6.7.0
commonmark==0.9.1
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
docopt==0.6.2
einops==0.3.2
fastapi==0.92.0
ffmpeg-python==0.2.0
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
google-auth==2.16.1
google-auth-oauthlib==0.4.6
greenlet==2.0.2
grpcio==1.51.3
h11==0.14.0
hmmlearn==0.2.8
huggingface-hub==0.12.1
HyperPyYAML==1.1.0
idna==3.4
joblib==1.2.0
julius==0.2.7
kiwisolver==1.4.4
librosa==0.9.2
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
more-itertools==9.0.0
mpmath==1.2.1
multidict==6.0.4
networkx==2.8.8
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
openai-whisper @ git+https://github.com/openai/whisper.git@7858aa9c08d98f75575035ecd6481f462d66ca27
optuna==3.1.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pooch==1.6.0
primePy==1.3
protobuf==3.20.1
pyannote.audio==2.1.1
pyannote.core==4.5
pyannote.database==4.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.5
pyDeprecate==0.3.2
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.5
pytorch-lightning==1.6.5
pytorch-metric-learning==1.7.3
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.4.2
rich==12.6.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.1
scipy==1.10.1
semver==2.13.0
sentencepiece==0.1.97
shellingham==1.5.0.post1
simplejson==3.18.3
singledispatchmethod==1.0
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
speechbrain==0.5.13
SQLAlchemy==2.0.4
starlette==0.25.0
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.2
torchaudio==0.13.1
torchmetrics==0.11.1
tqdm==4.64.1
transformers==4.26.1
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
Werkzeug==2.2.3
yarl==1.8.2
